Cleaned up Typescript API a bit
